+++ components/Header/BurgerBtn/BurgerBtn.module.scss
@@ -43,3 +43,4 @@
 .hidden {
   display: none;
 }
+
