+++ components/Header/Logo/LogoSvg/LogoSvg.jsx
@@ -5,8 +5,8 @@ const LogoSvg = ({ isFixed }) => {
     <svg
       className={isFixed ? css.hidden : css.visible}
       fill={"white"}
-      // width={"150px"}
-      // height={"75px"}
+      width={"150px"}
+      height={"75px"}
     >
       <use href={"/logo-sprite.svg#logo"} />
     </svg>
