+++ components/HomePageComponents/AboutStatisticSection/AboutStatisticSection.jsx
@@ -33,7 +33,7 @@ const AboutStatisticSection = () => {
 <ul className={css["topLineStatList"]}>
             <li className={css["counter__item"]}>
               <div className={css["counter__item__text"]}>
-                <h2 className={css["counter__num"]}><CountUp end={79} duration={3.4} enableScrollSpy={true}/></h2>
+                <h2 className={css["counter__num"]}><CountUp end={127} duration={3.4} enableScrollSpy={true}/></h2>
                 <p className={css["counter__descr"]}>Happy clients</p>
               </div>
             </li>
@@ -50,7 +50,7 @@ const AboutStatisticSection = () => {
             
             <li className={css["counter__BottomItem"]}>
               <div className={css["counter__item__text"]}>
-                <h2 className={css["counter__num"]}><CountUp end={1490} duration={2} enableScrollSpy={true}/></h2>
+                <h2 className={css["counter__num"]}><CountUp end={4479} duration={2} enableScrollSpy={true}/></h2>
                 <p className={css["counter__descr"]}>Completed Projects</p>
               </div>
             </li>
