+++ components/HomePageComponents/AboutStatisticSection/AboutStatisticSection.module.scss
@@ -1,17 +1,3 @@
-// .counter {
-//     background: #100028;
-
-//     @media screen and (min-width: 480px) and (max-width: 767px) {}
-
-//     @media screen and (min-width: 768px) and (max-width: 1023px) {}
-
-//     @media screen and (min-width: 1024px) {
-//         height: 500px;
-//         padding: 25px 25px;
-//         overflow: hidden;
-//     }
-// }
-
 .counter__content {
 
 
@@ -38,7 +24,7 @@
 .statList{
     display: flex;
     flex-direction: column;
-    gap:30px;
+    gap:20px;
 
 
     @media screen and (min-width: 768px){
@@ -51,7 +37,7 @@
 .topLineStatList{
     display: flex;
     flex-direction: column;
-    gap: 30px;
+    gap: 20px;
 
 
     @media screen and (min-width: 768px){
@@ -67,12 +53,11 @@ color: black;
 
 
 border-color: black;
-box-shadow: 6px 6px 18px 0px rgba(0, 0, 0, 0.3);
+box-shadow: 3px 3px 9px 0px rgba(0, 0, 0, 0.3);
 border: 1px solid black;
 
 border-radius: 16px;
 width: 100%;
-padding: 44px 24px 44px 24px;
 max-width: 450px;
 padding: 44px 24px 44px 24px;
 
@@ -95,13 +80,14 @@ padding: 44px 24px 44px 24px;
 
 
     border-color: black;
-    box-shadow: 6px 6px 18px 0px rgba(0, 0, 0, 0.3);
+    box-shadow: 6px 6px 9px 0px rgba(0, 0, 0, 0.3);
     border: 1px solid black;
 
 
     border-radius: 16px;
     max-width: 450px;
     padding: 44px 24px 44px 24px;
+
 
 
     @media screen and (min-width: 768px){
@@ -112,7 +98,7 @@ padding: 44px 24px 44px 24px;
 
     @media screen and (min-width: 1200px){
         padding: 44px 24px 44px 24px;
-        min-width: 1170px;
+        min-width: 1160px;
 
     }
 
@@ -121,138 +107,25 @@ padding: 44px 24px 44px 24px;
 
 
 .counter__num{
-    font-weight: 700;
-    font-size: 56px;
+
+
+    font-size: 60px;
+    font-weight: 500;
+    color: var(--secondaryTextColor);
+
+    letter-spacing: 2px;
+  
+    text-transform: uppercase;
+
+
     line-height: 70px;
 }
 
 
 .counter__descr{
-    font-weight: 400;
-    font-size: 24px;
-    line-height: 40px;
+    color: black;
+    font-size: 20px;
+    font-weight: 500;
+    line-height: calc(27 / 14);
 }
 
-// .row {
-//     display: flex;
-//     flex-wrap: wrap;
-
-//     list-style: none;
-
-//     @media screen and (min-width: 480px) {
-//         gap: 30px;
-
-//         padding-right: 15px;
-//         padding-left: 15px;
-//     }
-
-//     @media screen and (min-width: 1024px) {
-//         gap: 0;
-
-//         padding-right: 0;
-//         padding-left: 0;
-//     }
-// }
-
-// .counter__item {
-//     display: flex;
-//     align-items: center;
-//     justify-content: center;
-
-//     width: 100%;
-//     height: 255px;
-
-//     background: #171717;
-//     backdrop-filter: blur(15px);
-
-//     text-align: center;
-
-//     @media screen and (min-width: 576px) and (max-width: 767px) {
-//         width: 240px;
-//     }
-
-//     @media screen and (min-width: 768px) {
-//         width: 330px;
-//     }
-
-//     @media screen and (min-width: 1024px) {
-//         position: relative;
-
-//         transform: rotate(45deg);
-//         z-index: 1;
-        
-//         width: 255px;
-//     }
-// }
-
-// .counter__item:nth-child(1) {
-
-//     @media screen and (min-width: 1024px) {
-//         top: 208px;
-//     }
-    
-// }
-
-// .counter__item:nth-child(2) {
-
-//     @media screen and (min-width: 1024px) {
-//         right: 65px;
-//     }
-    
-// }
-
-// .counter__item:nth-child(3) {
-
-//     @media screen and (min-width: 1024px) {
-//         top: 208px;
-//         right: 130px;
-//     }
-    
-// }
-
-// .counter__item:nth-child(4) {
-
-//     @media screen and (min-width: 1024px) {
-//         bottom: 255px;
-//         left: 570px;
-//     }
-    
-// }
-
-// .counter__item__text {
-
-//     @media screen and (min-width: 1024px) {
-//         transform: rotate(-45deg);
-//     }
-    
-// }
-
-// .counter__image {
-//     display: block;
-
-//     max-width: 100%;
-
-//     margin-right: auto;
-//     margin-left: auto;   
-// }
-
-// .counter__item__text h2 {
-//     margin-bottom: 6px;
-//     margin-top: 18px;
-
-//     font-family: sans-serif;
-//     font-size: 60px;
-//     color: #ffffff;
-//     font-weight: 700;
-// }
-
-// .counter__item__text p {
-//     color: #ffffff;
-//     margin-bottom: 0;
-// }
-
-// .icon{
-    
-//     svg{width: 15px;
-//     height: 15px;}
-// }
