+++ components/HomePageComponents/FeedbackSection/FeedbackItem/FeedbackItem.module.scss
@@ -1,20 +1,25 @@
 .text {
-  padding: 32px;
-  border: 1px solid var(--secondaryTextColor);
+  border-color: black;
+  box-shadow: 3px 3px 9px 0px rgba(0, 0, 0, 0.3);
+  border: 1px solid black;
+  box-sizing: border-box;
+
   border-radius: 16px;
   // font-style: italic;
-  height: 230px;
+  min-height: 200px;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
   color: var(--secondaryTextColor);
+  padding: 16px 16px 16px 16px;
 
   @media screen and (min-width: 768px) {
-    height: 250px;
+    // height: 250px;
+    padding: 24px 24px 24px 24px;
   }
 
   & p {
-    font-size: 16px;
+    font-size: 18px;
     font-weight: 400;
     line-height: 1.14;
     padding: 0;
@@ -34,8 +39,8 @@
     @media screen and (min-width: 768px) {
       font-size: 26px;
     }
-    @media screen and (min-width: 1200px) {
-      font-size: 32px;
-    }
+    // @media screen and (min-width: 1200px) {
+    //   font-size: 20px;
+    // }
   }
 }
