+++ components/HomePageComponents/FeedbackSection/FeedbackSection.module.scss
@@ -20,3 +20,13 @@
   // padding: 0px 0 60px;
   position: relative;
 }
+
+
+.page_description_text {
+  text-align: center;
+  color: black;
+  font-size: 20px;
+  font-weight: 500;
+  line-height: calc(27 / 14);
+  margin-bottom: 20px;
+}
