+++ components/HomePageComponents/FeedbackSection/FeedbackSlider/FeedbackSlider.module.scss
@@ -1,3 +1,5 @@
+
+
 .controls {
   display: flex;
   align-items: center;
@@ -72,3 +74,4 @@ button.prev {
     display: flex;
   }
 }
+
