+++ components/HomePageComponents/FeedbackSection/LeaveFeedback/LeaveFeedback.jsx
@@ -5,15 +5,12 @@ import { FaCommentDots } from "react-icons/fa";
 export default function LeaveFeedback() {
   return (
     <>
-      <Container>
         <div className={css.wrapper}>
-          <h2>Here you can leave your feedback.</h2>
           <button type="button" className={css.button}>
-            Text us
+          LEAVE A COMMENT
             <FaCommentDots />
           </button>
         </div>
-      </Container>
     </>
   );
 }
