+++ components/HomePageComponents/FeedbackSection/LeaveFeedback/LeaveFeedback.module.scss
@@ -1,10 +1,12 @@
 .wrapper {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 20px;
+margin-left: auto;
+margin-right: auto;
+  margin-top: 28px;
   position: relative;
   z-index: 2;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 
   h2 {
     font-size: 18px;
@@ -22,7 +24,7 @@
   --af-white: #f3f3f3;
   --ch-white: #e1e1e1;
   border-radius: 8px;
-  max-width: 150px;
+  width: 200px;
   height: 44px;
   display: flex;
   align-items: center;
@@ -44,7 +46,7 @@
   }
 
   @media screen and (min-width: 768px) {
-    font-size: 18px;
+    font-size: 14px;
     padding: 16px 20px 16px 20px;
     svg {
       display: inline-block;
