+++ components/HomePageComponents/HomePage.module.scss
@@ -40,8 +40,7 @@
 
   
   color: black;
-  font-size: 24px;
+  font-size: 20px;
   font-weight: 500;
-  line-height: 36.77px;
-  // line-height: calc(27 / 14);
+  line-height: calc(27 / 14);
 }
