--- components/HomePageComponents/MainHero/ModalContactFormByHeroBtn/ModalContactFormByHeroBtn.module.scss
@@ -1,259 +1,0 @@
-.backdrop {
-    position: absolute;
-    z-index: 999;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: rgba(0,0,0,0.8);
-}
-
-.modal {
-
-width: 1200px;
-    position: absolute;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    top: 55%;
-    left: 50%;
-    border-radius: 15px;
-    transform: translateX(-50%) translateY(-50%);
-    // background-color: white;
-}
-
-
-.buttonCloseWrapper{
-    z-index: 100;
-    position: absolute;
-    top: 1%;
-    right: 38%;
-    border: 0;
-    width: 20px;
-    height: 20px;
-    background-color: transparent;
-
-
-    @media screen and (min-width: 768px) {
-      top: 1%;
-      right: 20%;
-    }
-  
-    @media screen and (min-width: 1200px) {
-      top: 1%;
-      right: 2%;
-    }
-
-
-
-}
-
-
-.buttonClose{
-    background-color: transparent;
-    border: 0;
-}
-
-.form {
-  padding: 30px 10px 30px 10px;
-    // text-align: center;
-    justify-content: center;
-    border: 1px solid black;
-    border-radius: 16px;
-    box-sizing: border-box;
-    box-shadow: 6px 6px 18px 0px rgba(0, 0, 0, 0.3);
-    color: #000000;
-    background-color: #ffffff;
-    display: flex;
-    justify-content: center;
-  flex-direction: column;
-    padding-top: 30px;
-    padding-bottom: 30px;
-    padding: 60px 40px 40px;
-    margin-left: auto;
-    margin-right: auto;
-    width: auto;
-  
-    @media screen and (min-width: 768px) {
-      display: flex;
-      flex-wrap: wrap;
-      max-width: 768px;
-      flex-direction: unset;
-    }
-  
-    @media screen and (min-width: 1200px) {
-      display: flex;
-      flex-wrap: wrap;
-      max-width: 1200px;
-      flex-direction: unset;
-    }
-  }
-  
-  .input-thumb {
-    margin-bottom: 20px;
-    height: 50px;
-    width: 100%;
-  
-  
-    @media screen and (min-width: 768px) {
-      height: 50px;
-      width: 47%;
-      margin-bottom: 20px;
-  
-      &:nth-child(2n) {
-        margin-left: auto;
-      }
-    }
-  
-    @media screen and (min-width: 1200px) {
-      height: 50px;
-      width: 47%;
-      margin-bottom: 40px;
-  
-      &:nth-child(2n) {
-        margin-left: auto;
-      }
-    }
-  }
-  
-  .input {
-    padding: 12px 15px;
-    // border: 1px solid black;
-    border: 0;
-    border-radius: 6px;
-    background-color: #f2f2f2;
-    width: 100%;
-  }
-  
-  .textarea-thumb {
-    width: 100%;
-    height: 320px;
-  
-    @media screen and (min-width: 1200px) {
-      width: 100%;
-      height: 320px;
-    }
-  }
-  
-  .text-area {
-    width: 250px;
-    height: 320px;
-    padding: 12px 15px;
-    // border: 1px solid black;
-    border: 0;
-    border-radius: 6px;
-    background-color: #f2f2f2;
-
-
-    @media screen and (min-width: 768px) {
-      width: 685px;
-    }
-  
-    @media screen and (min-width: 1200px) {
-      width: 1120px;
-    }
-
-
-
-  }
-  
-  .error {
-    padding: 5px 10px;
-    font-size: 10px;
-    color: var(--errorColor);
-  }
-  
-  .button {
-    --ch-black: #141414;
-    --eer-black: #1b1b1b;
-    --night-rider: #2e2e2e;
-    --af-white: #f3f3f3;
-    --ch-white: #e1e1e1;
-    border-radius: 8px;
-    width: 250px;
-    height: 34px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 8px;
-    margin-top: 40px;
-    padding: 16px 20px 16px 20px;
-    border: none;
-    color: white;
-    position: relative;
-    cursor: pointer;
-    transition-duration: 0.2s;
-    background-color: rgba(33, 33, 33, 0.8);
-  
-    font-weight: 600;
-    font-family: var(--open_sans);
-    text-align: center;
-  
-  
-    font-size: 1.17em;
-    line-height: 22px;
-    letter-spacing: 0.1rem;
-  
-
-
-    &:hover,
-    &:focus {
-      &::after,
-      &::before {
-        animation: none;
-        background: rgba(33, 33, 33, 0.8);
-        box-shadow: 6px 6px 12px #c5c5c5, -6px -6px 12px #ffffff;
-      }
-    }
-  }
-  
-  .button:before,
-  .button:after {
-    content: "";
-    position: absolute;
-    left: -2px;
-    top: -2px;
-    border-radius: 10px;
-    transition: all 0.8s ease;
-    background: linear-gradient(
-      45deg,
-      var(--ch-black),
-      rgba(33, 33, 33, 0.8),
-      rgba(33, 33, 33, 0.8),
-      var(--ch-white),
-      // var(--night-rider),
-      rgba(33, 33, 33, 0.8),
-      // var(--ch-black),
-      var(--ch-black)
-    );
-    background-size: 400%;
-    width: calc(100% + 4px);
-    height: calc(100% + 4px);
-    z-index: -1;
-    animation: steam 20s linear infinite;
-  }
-  
-  @keyframes steam {
-    0% {
-      background-position: 0 0;
-    }
-  
-    50% {
-      background-position: 400% 0;
-    }
-  
-    100% {
-      background-position: 0 0;
-    }
-  }
-  
-  .button:after {
-    filter: blur(50px);
-  }
-  
-  
-  .buttonWrapper{
-    position: relative;
-    z-index: 2;
-  }
